requirements.txt
ollama==0.3.3
marker-pdf==1.5.2

llama-index-llms-ollama
llama_index

langchain_text_splitters
langchain-ollama
langchain_chroma

streamlit
humanize
PyPDF2